+++ tests/test_join_paths_second_is_empty.c
@@ -1,0 +1,20 @@
+#include "fsio.h"
+#include "test.h"
+#include <stdlib.h>
+
+
+void test_impl()
+{
+  char *path = fsio_join_paths("file.txt", "");
+
+  assert_string_equal(path, "file.txt");
+
+  free(path);
+}
+
+
+int main()
+{
+  test_run(test_impl);
+}
+